requirements.txt
fastapi
uvicorn
python-dotenv
pypdf
google-generativeai
